Time: 38 ms (57.35%) | Memory: 7.9 MB (66.51%) - LeetSync

+++ leetcode/1457-minimum-difficulty-of-a-job-schedule/minimum-difficulty-of-a-job-schedule.cpp
@@ -5,18 +5,18 @@
 class Solution {
 public:
     int minDifficulty(std::vector<int>& jobDifficulty, int d) {
-        int32_t n = jobDifficulty.size();
+        int n = jobDifficulty.size();
         if (d > n) return -1;
 
-        const int32_t max = std::numeric_limits<int32_t>::max();
-        std::vector<int32_t> dp(n + 1, max);
+        const int max = std::numeric_limits<int>::max();
+        std::vector<int> dp(n + 1, max); // O(n) aux 
         dp[n] = 0;
 
-        for (int32_t day = 1; day <= d; ++day) {
-            std::vector<int32_t> next_dp(n + 1, max);
-            for (int32_t i = 0; i <= n - day; ++i) {
-                int32_t maxDiff = 0;
-                for (int j = i; j < n - day + 1; ++j) {
+        for (size_t day = 1; day <= d; ++day) { // O(n^2 * d)
+            std::vector<int> next_dp(n + 1, max);
+            for (size_t i = 0; i <= n - day; ++i) {
+                int maxDiff = 0;
+                for (size_t j = i; j < n - day + 1; ++j) {
                     maxDiff = std::max(maxDiff, jobDifficulty[j]);
                     if (dp[j + 1] != max) { 
                         next_dp[i] = std::min(next_dp[i], maxDiff + dp[j + 1]);
